backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
google-genai>=0.4.0
openai>=1.33.0
motor>=3.4.0
pymongo>=4.7.2
redis>=5.0.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
google-genai
